refactor: template 엔티티 구조 정리
