+++ src/entities/index.ts
@@ -2,4 +2,5 @@ export * from "./auth";
 export * from "./connection";
 export * from "./node";
 export * from "./oauth-token";
+export * from "./template";
 export * from "./workflow";
