+++ src/entities/template/api/get-template.api.ts
@@ -1,4 +1,4 @@
-import { request } from "../core";
+import { request } from "@/shared/api/core";
 
 import type { TemplateDetail } from "./types";
