+++ src/entities/template/api/instantiate-template.api.ts
@@ -1,4 +1,4 @@
-import { request } from "../core";
+import { request } from "@/shared/api/core";
 
 import type { InstantiateTemplateResponse } from "./types";
