+++ src/entities/template/model/index.ts
@@ -1,3 +1,4 @@
+export * from "./query-keys";
 export * from "./useCreateTemplateMutation";
 export * from "./useInstantiateTemplateMutation";
 export * from "./useTemplateListQuery";
