+++ src/entities/template/model/useTemplateListQuery.ts
@@ -1,8 +1,9 @@
 import { useQuery } from "@tanstack/react-query";
 
-import { templateApi } from "../../api";
-import { templateKeys } from "../../constants";
 import { type QueryPolicyOptions, toQueryMeta } from "@/shared/api";
+
+import { templateApi } from "../api";
+import { templateKeys } from "./query-keys";
 
 export const useTemplateListQuery = (
   category?: string,
