+++ src/shared/api/index.ts
@@ -1,5 +1,4 @@
 export * from "./client";
 export * from "./execution";
-export * from "./template";
 export * from "./workflow";
 export * from "./query-policy";
