+++ src/shared/model/index.ts
@@ -1,4 +1,3 @@
 export * from "./execution";
-export * from "./template";
 export * from "./workflow";
 export * from "./workflowStore";
